fix: findBoundingQuadの型をCV_32FC2→CV_32SC2に修正（四隅消失バグ）
orderPointsはdata32S(Int32Array)で座標を読むが、CV_32FC2(float)で
作成したMatはdata32Fに格納されるため、data32Sはゴミデータになっていた。
contourと同じCV_32SC2(int)にすることで正しく座標が読まれる。

+++ lib/opencv/paper-detection.ts
@@ -201,7 +201,7 @@ function findBoundingQuad(gray: Mat, cv: OpenCV): Mat | null {
   }
 
   // 外接矩形の4隅を Mat として返す
-  return cv.matFromArray(4, 1, cv.CV_32FC2, [minX, minY, maxX, minY, maxX, maxY, minX, maxY])
+  return cv.matFromArray(4, 1, cv.CV_32SC2, [minX, minY, maxX, minY, maxX, maxY, minX, maxY])
 }
 
 /**
